+++ CHANGELOG.md
@@ -1,5 +1,27 @@
 # Changelog
 
+## [Unreleased](https://github.com/zooniverse/markdownz/tree/master) (2023-10-30)
+- Refactor `Markdownz` as a functional component.
+- Refactor the Rehype code into `utils.getComponentTree`.
+- Return a React component tree from the `useMarkdownz` hook.
+
+```jsx
+// render HTML as JSX
+import { utils } from 'markdownz';
+const html = '<p>This is a test paragraph, with <a href="https://www.zooniverse.org">a link.</a>';
+const reactChildren = utils.getComponentTree({ html });
+return <div>{reactChildren}</div>;
+```
+
+```jsx
+import { useMarkdownz } from 'markdownz';
+
+const markdownChildren = useMarkdownz({ content: 'This is some markdown', debug: true });
+return <>{markdownChildren}</>;
+```
+
+**Full Changelog**: https://github.com/zooniverse/markdownz/compare/v9.0.0...master
+
 ## [v9.0.0](https://github.com/zooniverse/markdownz/tree/v9.0.0) (2023-10-30)
 Remove `dangerouslySetInnerHTML`. Render the HTML output with `rehype-react`.